+++ movie.re/src/components/App.tsx
@@ -1,6 +1,5 @@
 import { useState, useEffect, FC } from "react";
 import "../App.css";
-import Movie from "./Movie";
 import Search from "./Search";
 import Sort from "./Sort";
 
@@ -48,6 +47,7 @@ const App: FC = () => {
 
   const handleSort = () => {
       setSort(!sort);
+      
     };
   
 
@@ -57,16 +57,13 @@ const App: FC = () => {
       <h1>Movie Search Engine</h1>
       <Search search={ search } />
       <input type='checkbox' onChange={handleSort} checked={sort} />
-      <Sort movies={movies} sort={sort} />
       <div className="movies">
         {loading && !errorMessage ? (
           <span>loading...</span>
         ) : errorMessage ? (
           <div className="errorMessage">{errorMessage}</div>
         ) : (
-          movies.map((movie, index) => (
-            <Movie key={`${index}-${movie.Title}`} movie={movie} />
-          ))
+          <Sort movies={movies} sort={sort} />
         )}
       </div>
     </div>
