+++ movie.re/src/components/Debounce.tsx
@@ -1,0 +1,18 @@
+import { useState, useEffect } from "react";
+
+// debounce 関数を定義
+export const useDebounce = (value: string, delay: number) => {
+    const [debouncedValue, setDebouncedValue] = useState(value);
+
+    useEffect(() => {
+      const handler = setTimeout(() => {
+        setDebouncedValue(value);
+      }, delay);
+
+      return () => {
+        clearTimeout(handler);
+      };
+    }, [value, delay]);
+
+    return debouncedValue;
+  };
