+++ movie.re/src/components/Movie.tsx
