+++ movie.re/src/components/Sort.tsx
@@ -1,4 +1,5 @@
 import React, { useMemo } from 'react';
+import Movie from './Movie';
 
 interface SortProps {
   movies: MovieItem[];
@@ -21,13 +22,11 @@ const Sort: React.FC<SortProps> = ({ movies, sort }) => {
 
   return (
     <div className="sorted-movies">
-      {sortedMovies.map((movie, index) => (
-        <div key={index}>
-          <p>{movie.Title}</p>
-          <img src={movie.Poster} alt={movie.Title} />
-          <p>{movie.Year}</p>
-        </div>
-      ))}
+      <div className="movies">
+        {sortedMovies.map((movie, index) => (
+          <Movie key={index} movie={movie} />
+        ))}
+      </div>
     </div>
   );
 }
